fix check-dev
