+++ testkit/testkit.go
@@ -19,30 +19,24 @@ package testkit
 import (
 	"context"
 	"fmt"
-	"net"
 	"strings"
 	"sync"
 	"testing"
 	"time"
 
 	"github.com/pingcap/errors"
-	"github.com/pingcap/tidb/config"
-	"github.com/pingcap/tidb/domain"
 	"github.com/pingcap/tidb/expression"
 	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/parser/ast"
 	"github.com/pingcap/tidb/parser/terror"
-	"github.com/pingcap/tidb/server"
 	"github.com/pingcap/tidb/session"
 	"github.com/pingcap/tidb/sessionctx/variable"
-	"github.com/pingcap/tidb/util"
 	"github.com/pingcap/tidb/util/sqlexec"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/tikv/client-go/v2/tikv"
 	"github.com/tikv/client-go/v2/tikvrpc"
 	"go.uber.org/atomic"
-	"google.golang.org/grpc"
 )
 
 var testKitIDGenerator atomic.Uint64
@@ -497,20 +491,4 @@ func (c *RegionProperityClient) SendRequest(ctx context.Context, addr string, re
 		}
 	}
 	return c.Client.SendRequest(ctx, addr, req, timeout)
-}
-
-func SetUpRPCService(t *testing.T, addr string, dom *domain.Domain, sm util.SessionManager) (*grpc.Server, string) {
-	lis, err := net.Listen("tcp", addr)
-	require.NoError(t, err)
-	srv := server.NewRPCServer(config.GetGlobalConfig(), dom, nil)
-	port := lis.Addr().(*net.TCPAddr).Port
-	addr = fmt.Sprintf("127.0.0.1:%d", port)
-	go func() {
-		err = srv.Serve(lis)
-		require.NoError(t, err)
-	}()
-	config.UpdateGlobal(func(conf *config.Config) {
-		conf.Status.StatusPort = uint(port)
-	})
-	return srv, addr
 }
